src/app/src/requirements.txt

Flask
flask-cors
waitress
   
streamlit==1.33.0
langchain
langchain_community==0.2.7
oci==2.129.3
oracledb
pdfplumber
Pillow
pdf2image
pytesseract
pandas
requests
beautifulsoup4
python-dotenv
openpyxl
youtube-transcript-api
pdfminer.six
transformers
qdrant-client 
pypdf
